pyproject.toml
[tool.poetry]
authors = ["Daniel Stahl <danstahl1138@gmail.com>"]
description = "Utilities for cassandra and spark streaming specifically for streamstate"
homepage = "https://github.com/StreamState/streamstate-utils"
name = "streamstate-utils"
repository = "https://github.com/StreamState/streamstate-utils"
version = "0.0.0" #unused placeholder

[tool.poetry-dynamic-versioning]
enable = true
style = "semver"
vcs = "git"

[tool.poetry.dependencies]
marshmallow-dataclass = "^8.4.1"
python = "^3.7"
google-cloud-storage = "^1.37.1"

[tool.poetry.scripts]
run-test = "streamstate_utils.run_test:main"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
codecov = "^2.1.11"
coverage = {extras = ["toml"], version = "^5.5"}
mypy = "^0.812"
pyspark = "^3.1.1"
pytest = "^6.2.3"
pytest-cov = "^2.11.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["streamstate_utils"]

[tool.coverage.report]
show_missing = true
